requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.8.2
cffi==1.17.0
click==8.1.7
Flask==3.0.3
flask-talisman==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
Werkzeug==3.0.3
